Starting to refactor and generalize the router system
